test(core): split vite plugin adapter tests
