+++ packages/core/test/core/vite-plugin/helpers.ts
@@ -1,0 +1,20 @@
+export {
+  afterAll,
+  beforeAll,
+  build,
+  describe,
+  expect,
+  expectMarkerMappedToSourceLine,
+  expectPluginDiagnostic,
+  join,
+  lineOffsetMappings,
+  mkdtempSync,
+  readFileSync,
+  resolve,
+  rmSync,
+  test,
+  tmpdir,
+  vitePlugin,
+  writeFileSync,
+} from '../helpers';
+export type { RollupPlugin, ZntcPlugin } from '../helpers';
